requirements.txt
# Núcleo do projeto
opencv-python
opencv-contrib-python
mediapipe
numpy
Pillow
fastapi
uvicorn[standard]